Fix active item observation; fix tests
I think that we were missing the majority of active pane and pane item
changes before—we were only emitting the events when the active pane
container changed. With this commit, when the active container changes,
we observe the active pane; when the active pane changes, we observe
the active item.